Move razzia code to own app
