+++ razzia/templates/bread.html
@@ -1,6 +1,6 @@
-{% extends "admin/stregsystem/razzia/razzia_search.html" %}
+{% extends "admin/razzia/razzia_search.html" %}
 
-{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../../">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../../">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../">Brødrazzia</a></div>{% endblock %}
+{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../stregsystem/templates">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates/admin/stregsystem">Brødrazzia</a></div>{% endblock %}
 
 {% block member_present %}
 <div class="result">
