+++ razzia/templates/error_wizarderror.html
@@ -1,7 +1,7 @@
 ﻿{% extends "admin/base_site.html" %}
 
 {% block title %}Error when making a wizard!{% endblock %}
-{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../../">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../">Razzia</a>&nbsp;&rsaquo;&nbsp;Wizard</div>{% endblock %}
+{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates/admin/stregsystem">Razzia</a>&nbsp;&rsaquo;&nbsp;Wizard</div>{% endblock %}
 
 {% block content %}
 <center>
