+++ razzia/templates/members.html
@@ -1,7 +1,7 @@
 {% extends "admin/stregsystem/razzia/razzia.html" %}
 
 {% block title %}{{title|default:"Razzia"}}{% endblock %}
-{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../../../">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../../">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../">{{title|default:"Razzia"}}</a><a href=".">&nbsp;&rsaquo;&nbsp;Members</a></div>{% endblock %}
+{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../stregsystem">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates">Stregsystem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates/admin/stregsystem">{{title|default:"Razzia"}}</a><a href="../../stregsystem/templates/admin/stregsystem/razzia">&nbsp;&rsaquo;&nbsp;Members</a></div>{% endblock %}
 
 {% block razzia_content %}
 	<p><a href="../{{razzia.pk}}">Back to razzia</a></p>
