+++ razzia/templates/menu.html
@@ -2,7 +2,7 @@
 
 {% block title %}{{title|default:"Razzia"}}{% endblock %}
 
-{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../../">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../">Stregsystem</a>&nbsp;&rsaquo;&nbsp;{{title|default:"Razzia"}}</div>{% endblock %}
+{% block breadcrumbs %}<div class="breadcrumbs"><a href="../../stregsystem/templates">Hjem</a>&nbsp;&rsaquo;&nbsp;<a href="../../stregsystem/templates/admin">Stregsystem</a>&nbsp;&rsaquo;&nbsp;{{title|default:"Razzia"}}</div>{% endblock %}
 
 {% block razzia_content %}
 <a href="./new">
