+++ razzia/templates/razzia_search.html
@@ -5,7 +5,7 @@
 	<div class="titlebar">
 		<h1>{{title|default:"Razzia"}}!</h1>
 	</div>
-	<form action="./" method="post" class="content">{% csrf_token %}
+	<form action="../../stregsystem/templates/admin/stregsystem/razzia" method="post" class="content">{% csrf_token %}
 		<div class="grouped">
 			<span class="group-label fa fa-search">
 			</span>
