+++ razzia/templates/wizard.html
@@ -11,7 +11,7 @@
 
 <div id="content-main">
     <h1>Razzia wizard!</h1>
-    <form action="./" method="post" class="content">{% csrf_token %}
+    <form action="../../stregsystem/templates/admin/stregsystem/razzia" method="post" class="content">{% csrf_token %}
         <div>
             <table>
                 <tr>
