+++ razzia/templates/wizard_view.html
@@ -15,7 +15,7 @@
             <div class="titlebar">
                 <h1>{{ razzia_title }}</h1>
             </div>
-            <form action="./" method="get" class="content">
+            <form action="../../stregsystem/templates/admin/stregsystem/razzia" method="get" class="content">
                 <div class="grouped">
                     <span class="group-label fa fa-search">
                     </span>
